backend/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
supabase>=2.9.0
anthropic==0.28.0
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
